Revert "Channel last support (#423)"
This reverts commit d2b7713e4d9e91151e56d9f615233d350a1c6899.
