+++ nestedtensor/csrc/autograd_functions.cpp
@@ -43,11 +43,6 @@ Tensor NestedTensor_upsample_bilinear2d(
 Tensor NestedTensor_clone(
     const Tensor& src,
     c10::optional<c10::MemoryFormat> optional_memory_format) {
-  if (optional_memory_format) {
-    TORCH_CHECK(
-        *optional_memory_format == MemoryFormat::Contiguous, 
-        "Only contiguous format is unsupported by the NestedTensor_clone operator");
-  }
   return map_nested_tensor(
       [&optional_memory_format](Tensor a) {
         return at::clone(a, optional_memory_format);
@@ -125,85 +120,66 @@ Tensor NestedTensor_batch_norm(
       (var.dtype()     == torch::kHalf) &&
       (bias->dtype()   == torch::kHalf) &&
       (weight->dtype() == torch::kHalf) &&
-      get_is_cuda(input))
+      get_is_cuda(input)
+  )
   {
+  
     // Custom CUDA Half implementation.
     mean = mean.contiguous();
     Tensor bias_cont = (*bias).contiguous();
     Tensor weight_cont = (*weight).contiguous();
     Tensor running_var_cont = (*running_var).contiguous();
+  
+    Tensor output = input;
+    output = NestedTensor_contiguous(output);
+    Tensor input_buffer = get_buffer(output);
+    // Tensor output_buffer = input_buffer.clone();
+  
     auto self_opt_sizes = get_opt_sizes(input);
-
-    if (get_is_channel_last(input)) {
-      Tensor input_buffer = get_buffer_channel_last(input);
-      int64_t num_channel = weight_cont.size(0);
-      input_buffer = input_buffer.reshape({-1, num_channel});
-      at::Tensor invstd = at::rsqrt(running_var_cont + eps);
-      at::Tensor value = invstd * weight_cont;
-      at::Tensor value2 = -(mean * value - bias_cont);
-
-      input_buffer = at::addcmul(value2.reshape({1, num_channel}), input_buffer, value.reshape({1, num_channel}));
-      input_buffer = input_buffer.reshape(-1);
-
-      // Tensor output = input;
-      // output = output - mean.reshape(IntArrayRef(scalar_shape));
-      // output = output * invstd.reshape(IntArrayRef(scalar_shape));
-
-      // if (weight) {
-      //   output = output * weight->reshape(IntArrayRef(scalar_shape));
-      // }
-      // if (bias) {
-      //   output = output + bias->reshape(IntArrayRef(scalar_shape));
-      // }
-      // return output;
-      return wrap_buffer_channel_last(std::move(input_buffer), get_efficient_nested_size(input));
-    }
-
-    if (get_is_contiguous(input)) {
-      Tensor input_buffer = get_buffer(input);
-      Tensor nt_sizes_ =
-          get_efficient_nested_size(input).sizes(); // .to(torch::kInt32);
-      Tensor nt_sizes_1 = at::native::narrow(nt_sizes_, 1, 1, 1);
-      Tensor nt_sizes_2 = at::native::narrow(nt_sizes_, 1, 2, 1);
-      Tensor nt_sizes_all = nt_sizes_1 * nt_sizes_2;
-      int64_t* nt_sizes_all_ptr = nt_sizes_all.data_ptr<int64_t>();
-      at::Tensor numbers_t = at::empty({1 + (nt_sizes_all.size(0) * *self_opt_sizes[1])}, torch::kInt64);
-      int64_t* numbers_t_ptr = numbers_t.data_ptr<int64_t>();
-      numbers_t_ptr[0] = 0;
-      int64_t index = 1;
-      for (int64_t i = 0; i < nt_sizes_all.size(0); i++) {
-        for (int64_t j = 0; j < *self_opt_sizes[1]; j++) {
-          numbers_t_ptr[index] = (numbers_t_ptr[index - 1] + nt_sizes_all_ptr[i]);
-          index++;
-        }
+  
+    Tensor nt_sizes_ =
+        get_efficient_nested_size(input).sizes(); // .to(torch::kInt32);
+    Tensor nt_sizes_1 = at::native::narrow(nt_sizes_, 1, 1, 1);
+    Tensor nt_sizes_2 = at::native::narrow(nt_sizes_, 1, 2, 1);
+    Tensor nt_sizes_all = nt_sizes_1 * nt_sizes_2;
+    int64_t* nt_sizes_all_ptr = nt_sizes_all.data_ptr<int64_t>();
+    at::Tensor numbers_t = at::empty({1 + (nt_sizes_all.size(0) * *self_opt_sizes[1])}, torch::kInt64);
+    int64_t* numbers_t_ptr = numbers_t.data_ptr<int64_t>();
+    numbers_t_ptr[0] = 0;
+    int64_t index = 1;
+    for (int64_t i = 0; i < nt_sizes_all.size(0); i++) {
+      for (int64_t j = 0; j < *self_opt_sizes[1]; j++) {
+        numbers_t_ptr[index] = (numbers_t_ptr[index - 1] + nt_sizes_all_ptr[i]);
+        index++;
       }
-      Tensor nt_sizes = numbers_t.to(at::Device(kCUDA), torch::kInt32, true, true);
-    
-      c10::Half* mean_ptr = mean.data_ptr<c10::Half>();
-      c10::Half* running_var_ptr = running_var_cont.data_ptr<c10::Half>();
-      c10::Half* bias_ptr = bias_cont.data_ptr<c10::Half>();
-      c10::Half* weight_ptr = weight_cont.data_ptr<c10::Half>();
-    
-      at::cuda::CUDAStream defaultStream = at::cuda::getDefaultCUDAStream();
-      nested_tensor::cuda::batchnorm_inference_kernelLauncher(
-          input_buffer.data_ptr<c10::Half>(),
-          mean_ptr,
-          running_var_ptr,
-          c10::Half((float)(eps)),
-          weight_ptr,
-          bias_ptr,
-          input_buffer.data_ptr<c10::Half>(),
-          (int)(*self_opt_sizes[0]),
-          (int)(weight_cont.size(0)),
-          (int)(*self_opt_sizes[0] *
-                *self_opt_sizes[1] *
-                *self_opt_sizes[2] *
-                *self_opt_sizes[3]),
-          nt_sizes.data_ptr<int>(),
-          defaultStream
-          );
-      return wrap_buffer(std::move(input_buffer), get_efficient_nested_size(input), get_efficient_nested_stride(input));
     }
+    Tensor nt_sizes = numbers_t.to(at::Device(kCUDA), torch::kInt32, true, true);
+  
+    c10::Half* mean_ptr = mean.data_ptr<c10::Half>();
+    c10::Half* running_var_ptr = running_var_cont.data_ptr<c10::Half>();
+    c10::Half* bias_ptr = bias_cont.data_ptr<c10::Half>();
+    c10::Half* weight_ptr = weight_cont.data_ptr<c10::Half>();
+  
+    at::cuda::CUDAStream defaultStream = at::cuda::getDefaultCUDAStream();
+    nested_tensor::cuda::batchnorm_inference_kernelLauncher(
+        input_buffer.data_ptr<c10::Half>(),
+        mean_ptr,
+        running_var_ptr,
+        c10::Half((float)(eps)),
+        weight_ptr,
+        bias_ptr,
+        input_buffer.data_ptr<c10::Half>(),
+        // output_buffer.data_ptr<c10::Half>(),
+        (int)(*self_opt_sizes[0]),
+        (int)(weight_cont.size(0)),
+        (int)(*self_opt_sizes[0] *
+              *self_opt_sizes[1] *
+              *self_opt_sizes[2] *
+              *self_opt_sizes[3]),
+        nt_sizes.data_ptr<int>(),
+        defaultStream
+        );
+    return wrap_buffer(std::move(input_buffer), get_efficient_nested_size(output), get_efficient_nested_stride(output));
   }
 #endif
   auto scalar_shape = make_scalar_shape(get_dim(input), n_input);
