+++ nestedtensor/csrc/storage/EfficientSizeNode.h
@@ -12,7 +12,7 @@ inline at::Tensor stack_sizes(SizeNode size_node) {
   }
   std::vector<SizeNode> unbound_size_node = size_node.unbind();
   std::vector<int64_t> result_sizes_vector;
-  for(size_t i = 0; i < unbound_size_node.size(); i++) {
+  for(int64_t i = 0; i < unbound_size_node.size(); i++) {
     std::vector<int64_t> sizes = unbound_size_node[i].payload();
     if(i == 0) {
       result_sizes_vector.reserve(size_node.degree() * sizes.size());
@@ -21,7 +21,7 @@ inline at::Tensor stack_sizes(SizeNode size_node) {
       result_sizes_vector.push_back(sizes[j]);
     }
   }
-  return torch::tensor(result_sizes_vector, torch::kInt64).reshape({(int64_t)(size_node.degree()), -1});
+  return torch::tensor(result_sizes_vector, torch::kInt64).reshape({size_node.degree(), -1});
 }
 
 inline std::vector<c10::optional<int64_t>> construct_efficient_size(
@@ -196,20 +196,6 @@ inline EfficientSizeNode map_efficient_size(
     fn(sizes_ptr0 + i * sizes0.size(1), sizes_ptr1 + i * sizes1.size(1), sizes0.size(1));
   }
   return EfficientSizeNode(size_node0.height(), size_node0.structure(), sizes0);
-}
-
-template <class F>
-inline void apply_efficient_size(
-    F&& fn,
-    EfficientSizeNode& size_node0) {
-  at::Tensor sizes0 = size_node0.sizes();
-  int64_t* sizes0_ptr = sizes0.data_ptr<int64_t>();
-  int64_t structure0 = size_node0.structure();
-  for (int64_t i = 0; i < sizes0.size(0); i++) {
-    fn(sizes0_ptr + i * sizes0.size(1),
-       sizes0.size(1));
-  }
-  size_node0.refresh_opt_sizes();
 }
 
 template <class F>
