--- nestedtensor/csrc/transpose.h
@@ -1,20 +1,0 @@
-#include <nestedtensor/csrc/nested_tensor_impl.h>
-#include <torch/extension.h>
-#include <torch/library.h>
-
-namespace at {
-
-Tensor transpose_buffer(
-        Tensor nt_sizes_,
-        Tensor input_buffer,
-        Tensor output_buffer);
-
-Tensor transpose_nhwc_nchw(Tensor input);
-
-Tensor transpose_nchw_nhwc(Tensor input);
-
-Tensor transpose_nhwc_nchw_out(Tensor input, Tensor output);
-
-Tensor transpose_nchw_nhwc_out(Tensor input, Tensor output);
-
-}
